// apps/admin/src/shared/api/generated/ja-korea/schemas/boardMemberCreateRequestRole.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * JA Korea Homepage Admin API — JA Korea subset
 * Filtered for Homepage Admin JA Korea Orval codegen.
 * OpenAPI spec version: v1
 */

export type BoardMemberCreateRequestRole = typeof BoardMemberCreateRequestRole[keyof typeof BoardMemberCreateRequestRole];


export const BoardMemberCreateRequestRole = {
  BOARD_CHAIR_PRESIDENT: 'BOARD_CHAIR_PRESIDENT',
  FIDUCIARY_BOARD_MEMBER: 'FIDUCIARY_BOARD_MEMBER',
  BOARD_CHAIR: 'BOARD_CHAIR',
} as const;