// apps/admin/src/shared/api/generated/ja-korea/schemas/boardMemberUpdateItemRole.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * JA Korea Homepage Admin API — JA Korea subset
 * Filtered for Homepage Admin JA Korea Orval codegen.
 * OpenAPI spec version: v1
 */

export type BoardMemberUpdateItemRole = typeof BoardMemberUpdateItemRole[keyof typeof BoardMemberUpdateItemRole];


export const BoardMemberUpdateItemRole = {
  BOARD_CHAIR_PRESIDENT: 'BOARD_CHAIR_PRESIDENT',
  FIDUCIARY_BOARD_MEMBER: 'FIDUCIARY_BOARD_MEMBER',
  BOARD_CHAIR: 'BOARD_CHAIR',
} as const;